+++ packages/_private/e2e-demo-test/src/AllComponents.spec.ts
@@ -1,5 +1,5 @@
 import { type Locator, expect, test } from '@playwright/test';
-import type { SynCombobox, SynOptgroup } from '@synergy-design-system/components';
+import type { SynCombobox, SynOptgroup, SynSelect } from '@synergy-design-system/components';
 import { AllComponentsPage } from './PageObjects/index.js';
 import {
   createTestCases,
@@ -63,7 +63,7 @@ createTestCases(({ name, port }) => {
 
         await expect(AllComponents.getLocator('comboboxContent')).toBeVisible();
 
-        const combobox = await AllComponents.getLocator('comboboxComponent');
+        const combobox = await AllComponents.getLocator('combobox797');
         // Check that the displayed value is the text content of the option
         const displayedValue = await combobox.evaluate((ele: SynCombobox) => ele.displayLabel);
 
@@ -77,7 +77,7 @@ createTestCases(({ name, port }) => {
 
         await expect(AllComponents.getLocator('comboboxContent')).toBeVisible();
 
-        const combobox = await AllComponents.getLocator('comboboxComponent');
+        const combobox = await AllComponents.getLocator('combobox797');
         await combobox.evaluate((ele: SynCombobox) => {
           // eslint-disable-next-line no-param-reassign
           ele.value = 'option-3';
@@ -95,7 +95,7 @@ createTestCases(({ name, port }) => {
 
         await expect(AllComponents.getLocator('comboboxContent')).toBeVisible();
 
-        const combobox = await AllComponents.getLocator('comboboxComponent');
+        const combobox = await AllComponents.getLocator('combobox797');
         await combobox.evaluate((ele: SynCombobox) => {
           // eslint-disable-next-line no-param-reassign
           ele.value = 'option-4';
@@ -120,6 +120,56 @@ createTestCases(({ name, port }) => {
         expect(updatedDisplayedValue).toEqual('Option 4');
       });
     }); // regression#797
+
+    test.describe('Regression#813', () => {
+      test('should show the text content of the option, when value was set initially via property binding and options added dynamically', async ({ page }) => {
+        const AllComponents = new AllComponentsPage(page, port);
+        await AllComponents.loadInitialPage();
+        await AllComponents.activateItem('comboboxLink');
+
+        await expect(AllComponents.getLocator('comboboxContent')).toBeVisible();
+
+        const combobox = await AllComponents.getLocator('combobox813Level');
+        // Check that the displayed value is the text content of the option
+        const displayedValue = await combobox.evaluate((ele: SynCombobox) => ele.displayLabel);
+
+        expect(displayedValue).toEqual('Intermediate');
+      });
+
+      test('should reset the value of the combobox in a form to the initially set value via property binding', async ({ page }) => {
+        const AllComponents = new AllComponentsPage(page, port);
+        await AllComponents.loadInitialPage();
+        await AllComponents.activateItem('comboboxLink');
+
+        await expect(AllComponents.getLocator('comboboxContent')).toBeVisible();
+
+        const combobox = await AllComponents.getLocator('combobox813Form');
+
+        await combobox.evaluate((ele: SynCombobox) => {
+          // eslint-disable-next-line no-param-reassign
+          ele.value = '';
+        });
+
+        await combobox.click();
+        const options = await AllComponents.getLocator('combobox813FormOptions');
+        await options.last().click();
+
+        const [value, displayedValue] = await combobox.evaluate(
+          (ele: SynCombobox) => [ele.value, ele.displayLabel],
+        );
+        expect(value).toEqual('option-3');
+        expect(displayedValue).toEqual('Option 3');
+
+        const reset = await AllComponents.getLocator('comboboxFormReset');
+        await reset.click();
+
+        const [resetValue, resetDisplayedValue] = await combobox.evaluate(
+          (ele: SynSelect) => [ele.value, ele.displayLabel],
+        );
+        expect(resetValue).toEqual('option-1');
+        expect(resetDisplayedValue).toEqual('Option 1');
+      });
+    }); // regression#813
   }); // </syn-combobox>
 
   test.describe(`${name}: <SynOptgroup /> ${port}`, () => {
@@ -164,6 +214,52 @@ createTestCases(({ name, port }) => {
       });
     }); // regression#815
   }); // </syn-optgroup>
+
+  test.describe(`${name}: <SynSelect /> ${port}`, () => {
+    test.describe('Regression#813', () => {
+      test('should show the text content of the option, when value was set initially via property binding and options added dynamically', async ({ page }) => {
+        const AllComponents = new AllComponentsPage(page, port);
+        await AllComponents.loadInitialPage();
+        await AllComponents.activateItem('selectLink');
+
+        await expect(AllComponents.getLocator('selectContent')).toBeVisible();
+
+        const select = await AllComponents.getLocator('selectLevel');
+        // Check that the displayed value is the text content of the option
+        const displayedValue = await select.evaluate((ele: SynSelect) => ele.displayLabel);
+
+        expect(displayedValue).toEqual('Intermediate');
+      });
+
+      test('should reset the value of the select in a form to the initially set value via property binding', async ({ page }) => {
+        const AllComponents = new AllComponentsPage(page, port);
+        await AllComponents.loadInitialPage();
+        await AllComponents.activateItem('selectLink');
+
+        await expect(AllComponents.getLocator('selectContent')).toBeVisible();
+
+        const select = await AllComponents.getLocator('selectForm');
+        await select.click();
+        const options = await AllComponents.getLocator('selectFormOptions');
+        await options.last().click();
+
+        const [value, displayedValue] = await select.evaluate(
+          (ele: SynSelect) => [ele.value, ele.displayLabel],
+        );
+        expect(value).toEqual('option-3');
+        expect(displayedValue).toEqual('Option 3');
+
+        const reset = await AllComponents.getLocator('selectFormReset');
+        await reset.click();
+
+        const [resetValue, resetDisplayedValue] = await select.evaluate(
+          (ele: SynSelect) => [ele.value, ele.displayLabel],
+        );
+        expect(resetValue).toEqual('option-1');
+        expect(resetDisplayedValue).toEqual('Option 1');
+      });
+    }); // regression#813
+  }); // </syn-select>
 
   test.describe(`${name}: <SynTabGroup /> ${port}`, () => {
     test.describe('Regression#757', () => {