+++ packages/_private/e2e-demo-test/src/test.selector.ts
@@ -10,8 +10,12 @@ const AllComponentSelectors = {
   alertLink: '#tab-Alert',
 
   // Combobox
-  comboboxComponent: '#tab-content-Combobox syn-combobox',
+  combobox797: '#tab-content-Combobox syn-combobox[data-testid="combobox-797"]',
+  combobox813Form: '#tab-content-Combobox syn-combobox[data-testid="combobox-form-813"]',
+  combobox813FormOptions: '#tab-content-Combobox syn-combobox[data-testid="combobox-form-813"] syn-option',
+  combobox813Level: '#tab-content-Combobox syn-combobox[data-testid="combobox-level-813"]',
   comboboxContent: '#tab-content-Combobox',
+  comboboxFormReset: '#tab-content-Combobox form syn-button',
   comboboxLink: '#tab-Combobox',
 
   // Optgroup
@@ -22,6 +26,14 @@ const AllComponentSelectors = {
   optgroupSecondItem: '#tab-content-OptGroup syn-select syn-optgroup:nth-of-type(2)',
   optgroupThirdEnabledItems: '#tab-content-OptGroup syn-select syn-optgroup:nth-of-type(3) syn-option:not([disabled])',
   optgroupThirdItem: '#tab-content-OptGroup syn-select syn-optgroup:nth-of-type(3)',
+
+  // Select
+  selectContent: '#tab-content-Select',
+  selectForm: '#tab-content-Select syn-select[data-testid="select-form-813"]',
+  selectFormOptions: '#tab-content-Select syn-select[data-testid="select-form-813"] syn-option',
+  selectFormReset: '#tab-content-Select form syn-button',
+  selectLevel: '#tab-content-Select syn-select[data-testid="select-level-813"]',
+  selectLink: '#tab-Select',
 
   // Tabgroup
   tabGroupCustom: '#tab-content-TabGroup syn-tab:nth-of-type(2)',