+++ packages/_private/vanilla-demo/src/AllComponentParts/Combobox.ts
@@ -1,9 +1,27 @@
 import { html } from 'lit';
 
-export const Combobox = () => html`
-  <syn-combobox value="option-2">
-    <syn-option value="option-1">Option 1</syn-option>
-    <syn-option value="option-2">Option 2</syn-option>
-    <syn-option value="option-3">Option 3</syn-option>
-  </syn-combobox>
-`;
+export const Combobox = (regressions: Array< () => void> = []) => {
+  regressions.forEach((regression) => {
+    regression();
+  });
+
+  return html`
+    <syn-combobox data-testid="combobox-797" value="option-2">
+      <syn-option value="option-1">Option 1</syn-option>
+      <syn-option value="option-2">Option 2</syn-option>
+      <syn-option value="option-3">Option 3</syn-option>
+    </syn-combobox>
+
+    <syn-combobox data-testid="combobox-level-813" label="Experience" help-text="Please tell us your skill level." .value=${'2'}>
+    </syn-combobox>
+
+    <form>
+      <syn-combobox data-testid="combobox-form-813" .value=${'option-1'}>
+        <syn-option value="option-1">Option 1</syn-option>
+        <syn-option value="option-2">Option 2</syn-option>
+        <syn-option value="option-3">Option 3</syn-option>
+      </syn-combobox>
+      <syn-button type="reset">Reset</syn-button>
+    </form>
+  `;
+};