+++ packages/components/scripts/vendorism/custom/alert.vendorism.js
@@ -31,7 +31,7 @@ const transformStyles = (path, originalContent) => {
 };
 
 const transformComponent = (path, originalContent) => {
-  const contentWithoutCountdown = replaceSections([
+  const content = replaceSections([
     // Begin remove countdown
     [
       "@property({ type: String, reflect: true }) countdown?: 'rtl' | 'ltr';",
@@ -44,17 +44,6 @@ const transformComponent = (path, originalContent) => {
     // End remove countdown.
   ], originalContent);
 
-  // #782: Hide the active element when the alert is hidden
-  const content = addSectionsAfter([
-    [
-      "import type { CSSResultGroup } from 'lit';",
-      "import { blurActiveElement } from '../../internal/closeActiveElement.js';",
-    ],
-    [
-      '// Hide',
-      '      blurActiveElement(this);',
-    ],
-  ], contentWithoutCountdown);
   return {
     content,
     path,