+++ packages/components/web-test-runner.config.js
@@ -33,13 +33,21 @@ export default {
     })
   ],
   browsers: [
-    playwrightLauncher({ product: 'chromium' }),
+    
+    // Use concurrency of 1 for webkit. It seems to happen the same problem like with firefox in the CI
+    // @see https://github.com/modernweb-dev/web/issues/2374
+    playwrightLauncher({ product: 'chromium', concurrency: process.env.CI ? 1 : undefined }),
+
     
     // Enable firefox, but use concurrency of 1
     // @see https://github.com/modernweb-dev/web/issues/2374
     playwrightLauncher({ product: 'firefox', concurrency: 1 }),
 
-    playwrightLauncher({ product: 'webkit' })
+    
+    // Use concurrency of 1 for webkit. It seems to happen the same problem like with firefox in the CI
+    // @see https://github.com/modernweb-dev/web/issues/2374
+    playwrightLauncher({ product: 'webkit', concurrency: process.env.CI ? 1 : undefined }),
+
   ],
   testRunnerHtml: testFramework => `
     <html lang="en-US">