+++ packages/components/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@synergy-design-system/components-v2.26.2](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.26.1...components/2.26.2) (2025-03-31)
+
+
+### Bug Fixes
+
+* 🐛  Upgrade packages to latest versions ([#830](https://github.com/synergy-design-system/synergy-design-system/issues/830)) ([f5fe54f](https://github.com/synergy-design-system/synergy-design-system/commit/f5fe54fb55732b1e3efdf3062d55ee517aef4e81))
+
+# [@synergy-design-system/components-v2.26.1](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.26.0...components/2.26.1) (2025-03-31)
+
+
+### Bug Fixes
+
+* 🐛 Combobox and select not showing selected value if value was initially set via property binding ([#822](https://github.com/synergy-design-system/synergy-design-system/issues/822)) ([bb06d87](https://github.com/synergy-design-system/synergy-design-system/commit/bb06d87e64467d3710acf74868a69372be1225aa))
+
 # [@synergy-design-system/components-v2.26.0](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.25.0...components/2.26.0) (2025-03-27)